+++ manager/internal/manager/plex_runtime_other.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package manager
+
+func platformPlexToken() string {
+	return ""
+}